Fixed ClassCast exception when MentionsEditText is used directly
If we use MentionsEditText directly instead of RichEditorView and click
on the inserted span, the app crashes with ClassCast Exception.

Cause:
In the MentionSpan:onClick it was assumed that the parent will be always
RichEditorView. But this won't be the case if the MentionsEditText is
directly used. The parent can be any ViewGroup.

Fix:
In the MentionSpan:onClick, we don't need the reference of the parent.
We directly typecasted the Widget to MentionsEditText and called
updateSpan(), deselectAllSpan() on it.