Fixes to improve handling of mentions that are longer than one word
Previous commits fixed bugs when typing directly after a mention, but only handled cases where the mention was only a single word. The odd behaviors that were supposed to be fixed still existed for longer, multi-word mentions. This commit addresses these issues.

This commit also marks the release of version 1.2.1 of the library.