Fix #10 - Do not crash when using accessibility
Currently, any app crashes that uses our library when the user has accessibility enabled (or when trying to dump the view hierarchy, which uses the accessibility APIs). This is due to the accessibility framework trying to unparcel our custom MentionsEditable class using the default, system classloader (which does not know about our library's classes). To address this issue, this commit returns a normal Editable for accessibility purposes rather than a MentionsEditable, allowing it to be unparceled properly by the system.