Avoid ClassCastException in MentionsEditText#getMentionsText
Previously, the code in getMentionsText() cast the editable directly to be a MentionsEditable without checking the type. This can result in a ClassCastException if code using this view sets its text to be a CharSequence that is not a MentionsEditable. To avoid this, if the view is not showing a MentionsEditable, we will create and return a new MentionsEditable using the existing text.